+++ __tests__/auth.spec.js
@@ -32,17 +32,4 @@ describe('Authentication', () => {
       });
   });
 
-  it(`[POST /auth] login with using wallet id: ${seed.wallet.id}`, (done) => {
-    request(server)
-      .post('/auth')
-      .set('treetracker-api-key', seed.apiKey)
-      .send({ wallet: seed.wallet.id, password: seed.wallet.password })
-      .expect('Content-Type', /application\/json/)
-      .expect(200)
-      .end((err, res) => {
-        if (err) done(err);
-        expect(res.body).to.have.property('token');
-        done();
-      });
-  });
 });