+++ server/models/Wallet.spec.js
@@ -44,8 +44,8 @@ describe('Wallet Model', () => {
         error = e;
       }
 
-      expect(error.code).eql(403);
-      expect(error.message).eql(`The wallet 'wallet' already exists`);
+      expect(error.code).eql(409);
+      expect(error.message).eql(`The wallet "wallet" already exists`);
       expect(walletRepositoryStub.getByName).calledOnceWithExactly('wallet');
       expect(trustRepositoryStub.create).not.called;
       expect(walletRepositoryStub.create).not.called;