+++ server/repositories/TokenRepository.js
@@ -33,7 +33,7 @@ class TokenRepository extends BaseRepository {
   /*
    * select transaction table by transfer id, return matched tokens
    */
-  async getByTransferId(transferId, limit, offset = 0) {
+  async getByTransferId(transferId, limit, offset ) {
     return this._session.getDB().select('*')
         .from('token')
         .join('transaction', 'token.id', 'transaction.token_id')