+++ server/repositories/WalletRepository.js
@@ -14,11 +14,14 @@ class WalletRepository extends BaseRepository {
   }
 
   async getByName(wallet) {
-    try {
-      Joi.assert(wallet, Joi.string().pattern(/^\S+$/));
-    } catch (error) {
-      throw new HttpError(400, `invalid wallet name:${wallet}`);
-    }
+    Joi.assert(
+      wallet,
+      Joi.string()
+        .pattern(/^\S+$/)
+        .messages({
+          'string.pattern.base': `Invalid wallet name: "${wallet}"`,
+        }),
+    );
 
     const list = await this._session
       .getDB()
@@ -29,7 +32,10 @@ class WalletRepository extends BaseRepository {
     try {
       Joi.assert(list, Joi.array().required().length(1));
     } catch (error) {
-      throw new HttpError(404, `Could not find entity by wallet name: ${wallet}`);
+      throw new HttpError(
+        404,
+        `Could not find entity by wallet name: ${wallet}`,
+      );
     }
     return list[0];
   }
@@ -98,6 +104,12 @@ class WalletRepository extends BaseRepository {
 
     query = query.union(union1, union2);
 
+    // total count query (before applying limit and offset options)
+    const countQuery = this._session
+      .getDB()
+      .from(query.clone().as('q'))
+      .count('*');
+
     if (limitOptions && limitOptions.limit) {
       query = query.limit(limitOptions.limit);
     }
@@ -107,9 +119,9 @@ class WalletRepository extends BaseRepository {
     }
 
     const wallets = await query;
-    if (getCount) {
-      const count = await this._session.getDB().from(query.as('p')).count('*');
 
+    if (getCount) {
+      const count = await countQuery;
       return { wallets, count: +count[0].count };
     }