+++ server/repositories/WalletRepository.spec.js
@@ -92,7 +92,7 @@ describe('WalletRepository', () => {
         query.response([{ id: 1 }]);
       } else if (step === 2) {
         expect(query.sql).match(
-          /select.*wallet.*where.*actor_wallet_id.*request_type.*name.*limit.*/,
+          /select.*count.*where.*actor_wallet_id.*request_type.*name.*/,
         );
         query.response([{ count: 1 }]);
       }