+++ server/services/AuthService.spec.js
@@ -8,42 +8,42 @@ const WalletService = require('./WalletService');
 describe('AuthService', () => {
   it('signin', async () => {
     const walletObject = { salt: 'salt', password: 'hash' };
-    const getByIdOrNameStub = Sinon.stub(
+    const getByNameStub = Sinon.stub(
       WalletService.prototype,
-      'getByIdOrName',
+      'getByName',
     ).resolves(walletObject);
     const sha512Stub = Sinon.stub(HashService, 'sha512').returns('hash');
     const jwtSignStub = Sinon.stub(JWTService, 'sign').resolves('token');
     const details = { wallet: 'wallet', password: 'password' };
     const token = await AuthService.signIn(details);
-    expect(getByIdOrNameStub.calledOnceWithExactly(details.wallet)).eql(true);
+    expect(getByNameStub.calledOnceWithExactly(details.wallet)).eql(true);
     expect(sha512Stub.calledOnceWithExactly(details.password, 'salt')).eql(
       true,
     );
     expect(jwtSignStub.calledOnceWithExactly(walletObject)).eql(true);
     expect(token).eql('token');
-    getByIdOrNameStub.restore();
+    getByNameStub.restore();
     sha512Stub.restore();
     jwtSignStub.restore();
   });
 
   it('failed signin', async () => {
     const walletObject = { salt: 'salt', password: 'password' };
-    const getByIdOrNameStub = Sinon.stub(
+    const getByNameStub = Sinon.stub(
       WalletService.prototype,
-      'getByIdOrName',
+      'getByName',
     ).resolves(walletObject);
     const sha512Stub = Sinon.stub(HashService, 'sha512').returns('hash');
     const jwtSignStub = Sinon.stub(JWTService, 'sign').resolves('token');
     const details = { wallet: 'wallet', password: 'password' };
     const token = await AuthService.signIn(details);
-    expect(getByIdOrNameStub.calledOnceWithExactly(details.wallet)).eql(true);
+    expect(getByNameStub.calledOnceWithExactly(details.wallet)).eql(true);
     expect(sha512Stub.calledOnceWithExactly(details.password, 'salt')).eql(
       true,
     );
     expect(jwtSignStub.notCalled).eql(true);
     expect(token).eql(false);
-    getByIdOrNameStub.restore();
+    getByNameStub.restore();
     sha512Stub.restore();
     jwtSignStub.restore();
   });