+++ server/services/TrustService.js
@@ -14,9 +14,14 @@ class TrustService {
     state,
     type,
     request_type,
-    offset = 0,
+    offset,
     limit,
   }) {
+    // check if wallet exists first
+    // throws error if no wallet matching walletId exists
+    const walletService = new WalletService()
+    await walletService.getWallet(walletId)
+
     return this._trust.getTrustRelationships({
       walletId,
       state,